+++ jdk-1.7-parent/gmap3-parent/gmap3-examples/src/main/java/org/wicketstuff/examples/gmap/geocode/HomePage.java
@@ -14,7 +14,7 @@ import org.wicketstuff.gmap.api.GClientGeocoder;
 import org.wicketstuff.gmap.api.GInfoWindow;
 import org.wicketstuff.gmap.api.GLatLng;
 import org.wicketstuff.gmap.api.GMapType;
-import org.wicketstuff.gmap.geocoder.pojos.GeocoderStatus;
+import org.wicketstuff.gmap.geocoder.GeocoderStatus;
 
 /**
  * Demonstrates geocoding client- and serverwise.
