+++ jdk-1.7-parent/gmap3-parent/gmap3/src/main/java/org/wicketstuff/gmap/api/GClientGeocoder.java
@@ -10,7 +10,7 @@ import org.apache.wicket.request.Request;
 import org.apache.wicket.request.cycle.RequestCycle;
 import org.apache.wicket.util.time.Duration;
 import org.wicketstuff.gmap.GMapHeaderContributor;
-import org.wicketstuff.gmap.geocoder.pojos.GeocoderStatus;
+import org.wicketstuff.gmap.geocoder.GeocoderStatus;
 
 /**
  */
