+++ jdk-1.7-parent/gmap3-parent/gmap3/src/main/java/org/wicketstuff/gmap/geocoder/GeocoderException.java
@@ -17,8 +17,6 @@ package org.wicketstuff.gmap.geocoder;
 
 import java.io.IOException;
 
-import org.wicketstuff.gmap.geocoder.pojos.GeocoderStatus;
-
 public class GeocoderException extends IOException
 {
     private final GeocoderStatus _status;
