+++ jdk-1.7-parent/gmap3-parent/gmap3/src/main/java/org/wicketstuff/gmap/geocoder/GeocoderStatus.java
@@ -12,7 +12,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.wicketstuff.gmap.geocoder.pojos;
+package org.wicketstuff.gmap.geocoder;
 
 /**
  * Possible values of return status field.
