+++ jdk-1.7-parent/gmap3-parent/gmap3/src/main/java/org/wicketstuff/gmap/geocoder/pojos/GeocoderResult.java
@@ -15,6 +15,8 @@
  */
 package org.wicketstuff.gmap.geocoder.pojos;
 
+import org.wicketstuff.gmap.geocoder.GeocoderStatus;
+
 
 /**
  * {@link GeocoderResult} is the Resultset of the Google Geocode Request.<br/>
